Finish document upload side
