+++ src/Header2.js
@@ -6,43 +6,13 @@ import logo from './icons/logo.svg'
 import logout from './icons/logout.svg'
 import searchIcon from './icons/search.svg'
 
-export default function Header2(
+export default function Header2({
   showSearchIcon,
   checkInput,
-  searchedNumber,
-  active
-) {
+  searchedItem,
+  active,
+}) {
   const [showSearchBar, setShowSearchBar] = useState(false)
-  const Line = styled.div`
-    height: 45px;
-    border: solid 2px transparent;
-    border-bottom-color: rgb(201 193 171);
-    position: relative;
-    display: flex;
-    justify-content: center;
-  `
-  const Logo = styled.img`
-    z-index: -1;
-  `
-  const Logout = styled.img`
-    position: absolute;
-    top: 10px;
-    right: 25px;
-  `
-  const Search = styled.img`
-    position: absolute;
-    top: 10px;
-    left: 25px;
-  `
-  const Input = styled.input`
-    border: solid 2px rgb(201 193 171);
-    z-index: 1;
-    position: absolute;
-    left: 0;
-    top: 45px;
-    height: 25px;
-    width: 100%;
-  `
   function handleLogout() {
     fetch('/logout').then((window.location.href = `/`))
   }
@@ -50,7 +20,7 @@ export default function Header2(
     <Line>
       <Logout src={logout} onClick={handleLogout}></Logout>
       <Logo src={logo} alt="logo"></Logo>
-      {showSearchIcon.showSearchIcon === true ? (
+      {showSearchIcon === true ? (
         <Search
           src={searchIcon}
           onClick={() => setShowSearchBar(!showSearchBar)}
@@ -59,14 +29,16 @@ export default function Header2(
         ''
       )}
       {showSearchBar ? (
-        <Input
-          autoFocus
-          type="text"
-          placeholder="Suche nach Nummer oder Ort/Bauteil eingeben"
-          onChange={checkInput}
-          value={searchedNumber}
-          showSearchBar={active}
-        ></Input>
+        <Label onClick={() => setShowSearchBar(!showSearchBar)}>
+          <Input
+            autoFocus
+            type="text"
+            placeholder="Suche nach Dokumentenname eingeben"
+            onChange={checkInput}
+            value={searchedItem}
+            showSearchBar={active}
+          ></Input>
+        </Label>
       ) : (
         ''
       )}
@@ -76,3 +48,40 @@ export default function Header2(
 Header2.propTypes = {
   showSearchIcon: PropTypes.bool,
 }
+const Line = styled.div`
+  height: 45px;
+  border: solid 2px transparent;
+  border-bottom-color: rgb(201 193 171);
+  position: relative;
+  display: flex;
+  justify-content: center;
+`
+const Logo = styled.img`
+  z-index: -1;
+`
+const Logout = styled.img`
+  position: absolute;
+  top: 10px;
+  right: 25px;
+`
+const Search = styled.img`
+  position: absolute;
+  top: 10px;
+  left: 25px;
+`
+const Input = styled.input`
+  border: solid 2px rgb(201 193 171);
+  z-index: 1;
+  position: absolute;
+  left: 0;
+  top: 40px;
+  height: 25px;
+  width: 100%;
+`
+const Label = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 100px;
+  width: 100%;
+`
